Pull out codebase tests
